Updated to reflect recent v4.13 changes
v4.13: While the previous accuracy range was between 94% and 89%, the time limitation on Membean meant that you would get pretty consistent 90's and 88's. To counter this, I made the accuracy range far more unpredictable (78%-93%), which makes it seem as if the user is actively learning new words and mastering them. This change also has the upside of giving you less new words, which is nice. Finally, I restructured the way that the accuracy target is determined, because honestly the old method sucked ass.